+++ src/Processors/BrowserProcessor.js
@@ -14,12 +14,24 @@ import ProcessorBase from "./ProcessorBase";
 const BrowserProcessor = class extends ProcessorBase {
   /**
    * ProcessorBase ensures it is not being built outside a browser.
+   *
+   * @param {object} nav
+   *   window.Navigator.
+   * @param {object} win
+   *   window.
    */
-  constructor() {
+  constructor(nav, win) {
     super();
-    if (!navigator) {
-      throw new Error("BrowserProcessor is only usable on browser-run code.");
+    let actualNav = nav || (typeof navigator === "object" && navigator);
+
+    let actualWin = win || (typeof window === "object" && window);
+
+    if (typeof actualNav !== "object" || typeof actualWin !== "object" || !actualNav || !actualWin) {
+      throw new ReferenceError("BrowserProcessor is only usable on browser-run code.");
     }
+
+    this.navigator = actualNav;
+    this.window = actualWin;
   }
 
   /**
@@ -38,7 +50,6 @@ const BrowserProcessor = class extends ProcessorBase {
     const unknown = "unknown";
     const browserDefaults = {
       platform: unknown,
-      product: unknown,
       userAgent: unknown,
     };
 
@@ -48,11 +59,11 @@ const BrowserProcessor = class extends ProcessorBase {
     // Overwrite existing browser keys in context, keeping non-overwritten ones.
     for (const key in browserDefaults) {
       if (browserDefaults.hasOwnProperty(key)) {
-        result.browser[key] = navigator[key] ? navigator[key] : browserDefaults[key];
+        result.browser[key] = this.navigator[key] ? this.navigator[key] : browserDefaults[key];
       }
     }
 
-    result.browser.performance = (window.performance && window.performance.memory)
+    result.browser.performance = (this.window.performance && this.window.performance.memory)
       ? window.performance.memory
       : {};